requirements.txt
asgiref
Django
django-cors-headers
django-dotenv
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
tzdata
django-cors-headers
django-jazzmin  
